+++ HISTORY.rst
@@ -3,7 +3,6 @@
 History
 -------
 
-
 1.4.7 (2014-03-14)
 ~~~~~~~~~~~~~~~~~~
 
@@ -11,6 +10,8 @@ History
 - Bug fixing: max parallel resource creation check did not check that all the
               older tasks ended, only the last of the slot. This caused
               more tasks than permitted to be sent in parallel.
+- Improving multi-label training data uploads by zipping the extended file and
+  transforming booleans from True/False to 1/0.
 
 1.4.6 (2014-02-21)
 ~~~~~~~~~~~~~~~~~~